Revert "Move `OverridePackageCode` to `mokey-tests` crate"
This reverts commit dc0f310710194f276077deb75e36bb87be994bbb.
