+++ monkey-tests/src/lib.rs
@@ -22,7 +22,7 @@ use radix_engine::errors::{NativeRuntimeError, RuntimeError, VmError};
 use radix_engine::prelude::node_modules::ModuleConfig;
 use radix_engine::transaction::{TransactionOutcome, TransactionResult};
 use radix_engine::types::*;
-use radix_engine::vm::{NativeVmExtension, VmInvoke};
+use radix_engine::vm::OverridePackageCode;
 use radix_engine_interface::blueprints::package::PackageDefinition;
 use radix_engine_interface::blueprints::pool::{
     MultiResourcePoolInstantiateManifestInput, TwoResourcePoolInstantiateManifestInput,
@@ -809,41 +809,6 @@ impl<T: TxnFuzzer> FuzzTest<T> {
             self.cur_round = Round::of(1u64);
         } else {
             self.cur_round = Round::of(self.cur_round.number() + num_rounds);
-        }
-    }
-}
-
-#[derive(Clone)]
-pub struct OverridePackageCode<C: VmInvoke + Clone> {
-    custom_package_code_id: u64,
-    custom_invoke: C,
-}
-
-impl<C: VmInvoke + Clone> OverridePackageCode<C> {
-    pub fn new(custom_package_code_id: u64, custom_invoke: C) -> Self {
-        Self {
-            custom_package_code_id,
-            custom_invoke,
-        }
-    }
-}
-
-impl<C: VmInvoke + Clone> NativeVmExtension for OverridePackageCode<C> {
-    type Instance = C;
-
-    fn try_create_instance(&self, code: &[u8]) -> Option<C> {
-        let code_id = {
-            let code: [u8; 8] = match code.clone().try_into() {
-                Ok(code) => code,
-                Err(..) => return None,
-            };
-            u64::from_be_bytes(code)
-        };
-
-        if self.custom_package_code_id == code_id {
-            Some(self.custom_invoke.clone())
-        } else {
-            None
         }
     }
 }
