+++ monkey-tests/tests/fuzz_system.rs
@@ -1,4 +1,3 @@
-use monkey_tests::OverridePackageCode;
 use native_sdk::modules::metadata::Metadata;
 use native_sdk::modules::role_assignment::RoleAssignment;
 use radix_engine::errors::RuntimeError;
@@ -6,7 +5,7 @@ use radix_engine::kernel::kernel_api::{KernelNodeApi, KernelSubstateApi};
 use radix_engine::system::system_callback::SystemLockData;
 use radix_engine::transaction::TransactionReceipt;
 use radix_engine::types::*;
-use radix_engine::vm::VmInvoke;
+use radix_engine::vm::{OverridePackageCode, VmInvoke};
 use radix_engine_interface::api::{AttachedModuleId, ClientApi, LockFlags, ACTOR_STATE_SELF};
 use radix_engine_interface::blueprints::package::PackageDefinition;
 use rand::Rng;
